+++ src/Entities/Payments/Payment.php
@@ -22,21 +22,19 @@ class Payment implements JsonSerializable
     /** @var OrganisationShort */
     private $requestedByOrganisation;
     /** @var string */
-    private $assetTypeCode;
+    private $currencyCode;
     /** @var float */
     private $amount;
     /** @var string */
     private $paymentMethod;
     /** @var string */
     private $gateway;
-    /** @var integer */
-    private $grouping;
     /**
      * @var string
      */
     private $creationDate;
 
-    public function __construct($id, PaymentStatus $lastPaymentStatus, $gatewayTransactionId, $requestId, DisplayUser $requestedByUser, OrganisationShort $requestedByOrganisation, $assetTypeCode, $amount, $paymentMethod, $gateway, $grouping, $creationDate)
+    public function __construct($id, PaymentStatus $lastPaymentStatus, $gatewayTransactionId, $requestId, DisplayUser $requestedByUser, OrganisationShort $requestedByOrganisation, $currencyCode, $amount, $paymentMethod, $gateway, $creationDate)
     {
         $this->id = $id;
         $this->lastPaymentStatus = $lastPaymentStatus;
@@ -44,11 +42,10 @@ class Payment implements JsonSerializable
         $this->requestId = $requestId;
         $this->requestedByUser = $requestedByUser;
         $this->requestedByOrganisation = $requestedByOrganisation;
-        $this->assetTypeCode = $assetTypeCode;
+        $this->currencyCode = $currencyCode;
         $this->amount = $amount;
         $this->paymentMethod = $paymentMethod;
         $this->gateway = $gateway;
-        $this->grouping = $grouping;
         $this->creationDate = $creationDate;
     }
 
@@ -103,9 +100,9 @@ class Payment implements JsonSerializable
     /**
      * @return string
      */
-    public function getAssetTypeCode()
+    public function getCurrencyCode()
     {
-        return $this->assetTypeCode;
+        return $this->currencyCode;
     }
 
     /**
@@ -133,14 +130,6 @@ class Payment implements JsonSerializable
     }
 
     /**
-     * @return int
-     */
-    public function getGrouping()
-    {
-        return $this->grouping;
-    }
-
-    /**
      * @return string
      */
     public function getCreationDate()
@@ -160,11 +149,10 @@ class Payment implements JsonSerializable
             'requestId' => $this->getRequestId(),
             'requestedByUser' => $this->getRequestedByUser(),
             'requestedByOrganisation' => $this->getRequestedByOrganisation(),
-            'assetTypeCode' => $this->getAssetTypeCode(),
+            'currencyCode' => $this->getCurrencyCode(),
             'amount' => $this->getAmount(),
             'paymentMethod' => $this->getPaymentMethod(),
             'gateway' => $this->getGateway(),
-            'grouping' => $this->getGrouping(),
             'creationDate' => $this->getCreationDate(),
         ];
     }