+++ changelog.md
@@ -6,10 +6,22 @@ For now, please use all minor versions as major versions and each build version 
 
 ## [Unreleased]
 ## Added
+- `deliveryCost` to create order call
+
+### Changed
+- `assetTypeCode` to `currencyCode` on `Payment`
+
+### Removed
+- `groups` & `organisations` from `User`
+
+## [0.5.2]
+## Added
 - `JsonSerializable` to all entities - all entities can now be serialized to JSON
 - `markup` to `OrderItem`
 - `unitPrice` to `OrderItem`
 - `tip` to `Order`
+- `GetOrderInvoiceRequest` & `OrderInvoice`
+- `OrderDelivery` to `Order`
 
 ## [0.5.1]
 ### Fixed
@@ -65,7 +77,8 @@ For now, please use all minor versions as major versions and each build version 
 ## [< 0.3.1]
 ### History unavailable
 
-[Unreleased]: https://github.com/ordercloud/ordercloud-php/compare/0.5.1...HEAD
+[Unreleased]: https://github.com/ordercloud/ordercloud-php/compare/0.5.2...HEAD
+[0.5.2]: https://github.com/ordercloud/ordercloud-php/compare/0.5.1...0.5.2
 [0.5.1]: https://github.com/ordercloud/ordercloud-php/compare/0.5.0...0.5.1
 [0.5.0]: https://github.com/ordercloud/ordercloud-php/compare/0.4.2...0.5.0
 [0.4.2]: https://github.com/ordercloud/ordercloud-php/compare/0.4.1...0.4.2